PDF event flyer/backend/requirements.txt
awsebcli==3.7.4
base58==0.2.2
blessed==1.9.5
botocore==1.4.8
cement==2.4.0
colorama==0.3.3
Django==1.9.4
django-cors-headers==1.1.0
djangorestframework==3.3.3
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
futures==3.0.4
html5lib==0.9999999
jmespath==0.9.0
lambda-packages==0.3.0
mysql-connector-python-rf==2.1.3
MySQL-python==1.2.5
pathspec==0.3.3
Pillow==3.1.1
pyasn1==0.1.9
PyPDF2==1.25.1
python-dateutil==2.4.2
PyYAML==3.11
reportlab==3.3.0
requests==2.6.2
rsa==3.3
s3transfer==0.0.1
semantic-version==2.5.0
six==1.10.0
texttable==0.8.4
tqdm==3.7.1
wcwidth==0.1.6
websocket-client==0.35.0
Werkzeug==0.11.4
wheel==0.24.0
wsgi-request-logger==0.4.4
xhtml2pdf==0.0.6
xmltodict==0.10.1
zappa==0.13.1
